Refactor and enable better default playlist integration
The default playlist is queried for a random song when there is only one
song on the queue and that song is finishing. The song from the default
playlist is added to the queue as normal to integrate better with
playing songs. This way new songs requested by chat will simply go on
the queue and play after the default song is done (before the new song
would interrupt the old one).

Refactored some response parsing from spotify and got rid of a couple
dead functions.