Update song information if we're already playing the song
In some cases, we would restart the song or interrupt a previous song
that is playing because we would start playing right away, not knowing
what is currently playing. Now we check before we start playing the song
so we don't interrupt another song that's playing (if it's not already
tracked by the queue) and so we continue playing where we left off and
there is no hitch in the music.