Reorder start_tracking_playback and start_playing
If we start tracking playback first we have a race condition so we might
end up thinking the current song is 'done' because spotify is currently
playing a different song and we haven't ran our current 'start_playing'
function.

This was causing spotify to potentially skip several songs if it got
stuttered for a few seconds.